+++ uni_modules/hf-ui/index.js
@@ -1,5 +1,5 @@
 // 工具函数
-import * as utils from './libs/util/utils.js';
+// import * as utils from './libs/util/utils.js';
 
 // 设置获取数据的函数
 import { setUtils } from './libs/util/loadData.js';
@@ -15,6 +15,6 @@ export default {
 	install,
 	
 	setUtils,
-	...utils,
+	// ...utils,
 	// formMixin,
 }